requirements.txt
flask
ultralytics
Pillow
opencv-python
rmn 